added bash if statement.

+++ .platform/hooks/predeploy/custom-predeploy-script.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 sudo su
 cd /var/app/current
-mkdir node_modules
-npm install --omit=dev
+
+if [ ! -e "./node_modules" ]; then
+  mkdir node_modules
+  npm install --omit=dev
+fi
+
 echo "PREDEPLOY COMPLETED"
